Bug fixes and UI updates

+++ app/src/main/java/com/example/abhishektiwari/gpacalculator/UserGrades.java
@@ -1,12 +1,12 @@
 package com.example.abhishektiwari.gpacalculator;
 
 public class UserGrades {
-    public int sem1,sem2,sem3,sem4,sem5,sem6,sem7,sem8;
+    public double sem1,sem2,sem3,sem4,sem5,sem6,sem7,sem8;
 
     public UserGrades(){
 
     }
-    public UserGrades(int sem1,int sem2,int sem3,int sem4,int sem5,int sem6,int sem7,int sem8){
+    public UserGrades(double sem1,double sem2,double sem3,double sem4,double sem5,double sem6,double sem7,double sem8){
         this.sem1 = sem1;
         this.sem2 = sem2;
         this.sem3 = sem3;
@@ -21,7 +21,7 @@ public class UserGrades {
         return String.valueOf(sem1);
     }
 
-    public void setSem1(int sem1) {
+    public void setSem1(double sem1) {
         this.sem1 = sem1;
     }
 
@@ -29,7 +29,7 @@ public class UserGrades {
         return String.valueOf(sem2);
     }
 
-    public void setSem2(int sem2) {
+    public void setSem2(double sem2) {
         this.sem2 = sem2;
     }
 
@@ -37,7 +37,7 @@ public class UserGrades {
         return String.valueOf(sem3);
     }
 
-    public void setSem3(int sem3) {
+    public void setSem3(double sem3) {
         this.sem3 = sem3;
     }
 
@@ -45,7 +45,7 @@ public class UserGrades {
         return String.valueOf(sem4);
     }
 
-    public void setSem4(int sem4) {
+    public void setSem4(double sem4) {
         this.sem4 = sem4;
     }
 
@@ -53,7 +53,7 @@ public class UserGrades {
         return String.valueOf(sem5);
     }
 
-    public void setSem5(int sem5) {
+    public void setSem5(double sem5) {
         this.sem5 = sem5;
     }
 
@@ -61,7 +61,7 @@ public class UserGrades {
         return String.valueOf(sem6);
     }
 
-    public void setSem6(int sem6) {
+    public void setSem6(double sem6) {
         this.sem6 = sem6;
     }
 
@@ -69,7 +69,7 @@ public class UserGrades {
         return String.valueOf(sem7);
     }
 
-    public void setSem7(int sem7) {
+    public void setSem7(double sem7) {
         this.sem7 = sem7;
     }
 
@@ -77,7 +77,7 @@ public class UserGrades {
         return String.valueOf(sem8);
     }
 
-    public void setSem8(int sem8) {
+    public void setSem8(double sem8) {
         this.sem8 = sem8;
     }
 }
